Debug injecting yaw angel

+++ training/coach.py
@@ -81,7 +81,7 @@ class Coach:
 			for batch_idx, batch in enumerate(self.train_dataloader):
 				self.optimizer.zero_grad()
 				x, y, yaw = batch
-				x, y, yaw = x.to(self.device).float(), y.to(self.device).float(), yaw.to(self.device)
+				x, y, yaw = x.to(self.device).float(), y.to(self.device).float(), yaw.to(self.device).float()
 				y_hat, latent = self.net.forward(x, yaw, return_latents=True)
 				loss, loss_dict, id_logs = self.calc_loss(x, y, y_hat, latent)
 				loss.backward()
@@ -122,7 +122,7 @@ class Coach:
 			x, y, yaw = batch
 
 			with torch.no_grad():
-				x, y, yaw = x.to(self.device).float(), y.to(self.device).float(), yaw.to(self.device)
+				x, y, yaw = x.to(self.device).float(), y.to(self.device).float(), yaw.to(self.device).float()
 				y_hat, latent = self.net.forward(x, yaw, return_latents=True)
 				loss, cur_loss_dict, id_logs = self.calc_loss(x, y, y_hat, latent)
 			agg_loss_dict.append(cur_loss_dict)
